clinic/users/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0      
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
distlib==0.3.4
Django==4.0.2
django-bootstrap-v5==1.0.11 
django-bootstrap4==21.2     
django-environ==0.8.1       
django-extra-views==0.14.0  
django-filter==21.1
django-rest-authtoken==2.1.4
djangorestframework==3.13.1 
environ==1.0
filelock==3.4.2
Flask==2.0.3
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.0.3
